tests/functional: let check clean directories of passed tests in default
Some tests consume lots of disk space and cause disk full (at least on
my box with 64GB SSD). This patch lets the check script clean
directories of passed tests in default. In addition, this patch adds a
new option -keep-passed. When this option is passed, the script keeps
directories of passed tests. This would be useful when we need to
check correctness of test scriptes themselves.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>