sheep: use sd_hash instead of fnv_64a_buf
This patch uses sd_hash instead of fnv_64a_buf for better object
dispersion.

This also fixes the problem that we are using the number of disks for
calculating virtual disk ids.  Without this patch, virtual disk ids of
the existing disks will be changed for each we plug/unplug a disk; a
lot of object movement will happen.

Please note that this patch breaks backward compatibility because this
changes a hash function of our consistent hashing algorithm, which
means that object location will be also changed.  The upgrade code for
the existing users will be appeared before releasing the next sheepdog
release.

I've not changed some hash calculation which is used for getting
object file names (vdi object name and attribute object name).  It is
because changing them makes it very hard to implement the upgrade code
and forces us to modify qemu client codes too.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>