increase the default number of virtual nodes
We gave up backward compatibility in the previous patch, so it's a
good chance to increase the default number of virtual nodes.

The number means the default granularity of how much data will be
moved after node change events.  E.g., if the number of virtual nodes
is 5, at least one-fifth data will be moved when object movement
happens.

There is no good rational reason for the number 128.  The bigger
number requires more memory space.  I've chose it just because it's
the same number as the current default virtual disks.

Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>