sheep: use rbtree to manage struct sd_node
No algorithm nor functionality changes, but involes a lot of api changes that
assume a static array of sd_nodes.

With this patch, cluster driver will pass a rbtree of nodes directly to sheep
handlers and all the way down we'll play with rbtree instead of the old static
array.

I do not yet move static sd_nodes array in struct epoch_log and struct
cluster_info to dynamic structures. So basically we are still limited by max
node roof, but we are way close to get rid of it.

Shepherd driver is consdiered to be changed later because it might be redesigned
in the future and not stale for now.

Signed-off-by: Liu Yuan <namei.unix@gmail.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>