gateway: remove shortcut for local node in gateway_forward_request
Shortcut forward-reqeust to local node doesn't improve performance a lot because
we forward the requests to many nodes concurrently and wait for all the
responses of all the requests, slowest one determine the how long we take for
this request.

Removing the shortcut for local node doesn't reduce the performance but result
in a more simplifed code.

Signed-off-by: Liu Yuan <namei.unix@gmail.com>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>