dog: fix return value of vdi_backup()
"dog vdi backup" command returns 16 as its status code even if it
successfully ended.

This is because the current vdi_backup function returns the byte count
returned by the last xwrite call in the function.

This fixes the issue by overwriting the return code with EXIT_SUCCESS
at the end of the normal path of vdi_backup().

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>
Reviewed-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>