sheep: fix improper strncmp usage in fill_vdi_info_range()
The following usage of strncmp in fill_vdi_info_range function is a
misuse.  It is inefficient and also unsafe because it can overrun if
inode->name is not terminated with a null byte due to malformed inode
object:

  strncmp(inode->name, name, strlen(inode->name))

This fixes the issue by replacing the size argument
"strlen(inode->name)" with "sizeof(inode->name)".

Signed-off-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>