erasure: allow x nodes to serve write request with x:y policy
There is no technical reason that prvent us from serving the both read and write
requests if there are x nodes alove in the cluster with x:y policy.

E.g, with 4 nodes alive with 4:2 formated, the cluster will be still operational
instead of denial of service with this patch.

This patch also fixes a memory leak in erorr case.

Reviewed-by: Robin Dong <sanbai@taobao.com>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>