sheep: introduce strict mode for write
We make sure we write the exact number of copies to honor the promise of the
redundancy for "strict mode". This means that after writing of targeted data,
they are redundant as promised and can withstand the random node failures.

For example, with a 4:2 policy, we need at least write to 6 nodes with data
strip and parity strips. For non-strict mode, we allow to write successfully
only if the data are written fully with 4 nodes alive.

Reviewed-by: Robin Dong <sanbai@taobao.com>
Reviewed-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>