sheep: daemon should exit directly if there is no availabe disk
Currently, the sheep daemon will only report "multi-disk support need xattr feature"
and continue to run when all the data pathe are not support xattr.But if we create
a vdi, it will report path is broken.
So, we need to check the return code of md_add_disk() and exit if all the path is
unavailabe.

This patch also fix some log level in source code.

Signed-off-by: Robin Dong <sanbai@taobao.com>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>