sheep/http: make kv_delete_object more safe
For object create/delete, we can't easily maintain the bnode consistent by
playing around the operations order.

We should inform the user the deletion failure if bnode_update() fails even
though we might delete the onode successfully. Then subsequent 'delete' for
the same object won't skew up the bnode metadata.

The true fix for the inconsistency (for whatever reaons it happens), is a check
request that does a server side consistency check. This is left for a future
patch.

Another fix is that we drop the redundant data about bytes_used, object_counts
from bnode, and so for "HEAD" operation, we just iterate all the objects. This
can't scale if we have huge objects.

Reviewed-by: Robin Dong <sanbai@taobao.com>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>