dog: log dog operation in syslog
This patch is based on a request from our internal team. Currently,
sheepdog doesn't provide a way of logging admin's operation. The lack
of this feature is serious because we cannot find any evidence of
operation.

This patch adds the logging feature in dog command. If the environment
variable, SHEEPDOG_DOG_LOG, has value, dog command logs its command
line argument with syslog. If SHEEPDOG_DOG_LOG_PATH is specified
additionally, the log is written to the file of the path.

Below is an example of the log:
- with standard syslog
  Dec 27 10:57:42 dhcpff33 sheepdog admin operation[9809]:  dog
  cluster format -c 1
- with custom path
  2013 Dec 27 17:20:56 :  dog vdi list

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
[ remove extra blank lines - Yuan]
Signed-off-by: Liu Yuan <namei.unix@gmail.com>