dog: forbid invalid copy policy option of cluster and vdi
"dog vdi create" and "dog cluster format" of stable-0.7 don't treat a
command line option for copy policy like "3:1" as an invalid
thing. This would be confusing for users who want to use the feature
of erasure coding. This patch lets dog command cause error when such
an option is passed.

Below is an example of error messages:

$ dog cluster format -c 3:1
copy number 3:1 is invalid
If you want to use the erasure coding feature, please use sheepdog 0.8 or later
$ dog vdi create test 1G -c 3:X
copy number 3:X is invalid
If you want to use the erasure coding feature, please use sheepdog 0.8 or later

Cc: Marcin Miroslaw <marcin@mejor.pl>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>