sheep: exit when a passed path of log dir is not a directory
This patch lets sheep exit as soon as possible when the option of log
directory (specified with "-l dir=") is not a directory, but an
ordinal file. When a path of ordinal file is passed, logger process
runs as a defunct process. This behavior is confusing for
administrators.

The purpose of this change is preventing operation miss. If sheep is
passed a wrong path name of log directory, it should notify the
mistake to administrator. Running without correct logging is really
wrose in production environment.

In addition, this patch does a trivial cleaning. The variable path in
main() of sheep is too generic. So this patch renames it to log_path.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>