logger: add a unified log destination option and syslog support
This patch adds a new parameter "dst=" to the "-l" option for
specifying logging destination. Possible parameters are below:
- default: a dedicated file under sheep's directory
- syslog: standard syslog
- stdout: standard output

In addition, this patch removes "-o" option and "-f" option, previous
options for logging with stdout and let sheep run as a foreground
process. Now it is replaced with "-l dst=stdout".

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>