logger: add syslog support as a log destination
A new option '-s' is added for using syslog as a log destination.

This patch is based on the commit 275dcf916ed0: logger: add a unified
log destination option and syslog support. But the change for
backporting is too large so this commit log is rewrote as a new one.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>

Changes for backporting are reviewed by MORITA Kazutaka.
Reviewed-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>