sheep: changed end of auto-recovery message to notice
Sheep output the end of auto-recovery log message (all nodes are recovered)
at DEBUG level.
When sheep nodes are working in the INFO level or upper,
the operator could not get this message
even if operators want to know the completion of auto-recovery.

This patch raised the log message to the NOTICE level.

Signed-off-by: Yoshinori Matsuo <matsuo.yoshinori@lab.ntt.co.jp>
Reviewed-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: MORITA Kazutaka <morita.kazutaka@lab.ntt.co.jp>