sheep: fix memory leak in _invert_mat()
First run swift interface on sheepdog, thean kill a node and wait for
recovery. After it, the RSS of sheep daemon will increase to about 8GB
which is a sign of memory leak.

Then we use valgrind to check sheepdog daemon and finally find out the
leak point: _invert_mat() in lib/fec.c

Signed-off-by: Robin DOng <sanbai@taobao.com>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>