sheep: avoid dead loop when zookeeper session is expired
When zookeeper session is expired, zk_get_least_seq() will return
ZNONODE, but we have not check the return value of zk_get_least_seq()
so the routine will fall in dead loop.

To avoid it, we add check for return value of zk_get_least_seq() and
jump to recreate sequential-node if find out the ZNONODE event.

Signed-off-by: Robin Dong <sanbai@taobao.com>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>