util/sd_inode: structure and function renaming
There is actually no extent at all in the implementation btree, it is
traditional index and indirect index. So better name it as it is.

And refine some comments.

There are still some remnant variables that is extent related and hope we can
get rid of them completely for future refactor this sd_inode.c

Reviewed-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Reviewed-by: Robin Dong <sanbai@taobao.com>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>