sheep/cluster: delete znodes when leaving cluster
A sheep node call zk_lock() to create a znode in zookeeper and become the
owner of the lock, then if all disks under this sheep node is broken
suddenly, the node will have to leave cluster.Now, the znode will stay
in "/lock" and make other sheep nodes waiting it forever.

So we should delete znodes which created by this sheep node when leaving
cluster.

Signed-off-by: Robin Dong <sanbai@taobao.com>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>