dog/farm: add snap log header
We don't store any previous cluster state for farm, so we can't restore old state
like redundancy scheme.

This patch introduce a header for snap file and as a side effect, we add a magic
and version check capability for cluster snapshot.

This will break the compatibility of old cluster snapshot. For simplicity, we just
error out if we are operating on the old snapshot data.

Reviewed-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>