sheep/recovery: reduce network overhead of preparing object list
In the stage of preparing object list, every node tries to fetch
entire object lists from all nodes of the cluster, and then,
screens out objects that don't belong to it's own. This behavior
transfers quite a lot of unnecessary object id over the network.

This patch resolve the issue by:
1. receiver node prepare objects that belong to the sender node.
2. sender node don't screen out objects any longer because it is
   already done by the receiver, but merging object lists fetched
   from different nodes is essential.

Signed-off-by: Ruoyu <liangry@ucweb.com>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>