add Robin as a subsystem maintainer of HTTP interface
Robin introduced big contributions (e.g. hypervolume) to sheepdog and
is a heavy user of the feature of HTTP interface for object storage at
Taobao, inc. So this patch adds him to MAINTAINERS as a subsystem
maintainer of HTTP interface (sheep/http/*).

Of course, current HTTP interface subsystem isn't separated cleanly
from the core components of sheepdog. Changes related to the subsystem
will affect the core part often. But I want to increase productivity
of the community and it will be helpful if Robin can be a maintainer
of the subsystem even his territory is limited. I'd like to pull
changes of HTTP interface from Robin's repository directly.

Cc: Robin Dong <robin.k.dong@gmail.com>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>