dog/node: confirm before md plug and unplug
Once a directory is plugged into the node, all data in it will be
removed. Therefore, it is better to print a warning message to
user first for fear that he or she specifies a wrong path.

Unplugging also needs a confirmation before the execution because
recovery will impact the guest OS.

For admin's scripts, the option -f (--force) is introduced to avoid
the confirmation.

Test case 057, 063, 064 are also updated.

Signed-off-by: Ruoyu <liangry@ucweb.com>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>