optimize repairing replica in dog vdi check command
Current repairing replica logic is described as below.

1. Read data stream from majority node to dog process's buffer.
2. Write buffer to destinate node to finish repairing.

This behavior leads to two side-effects.

1. Data is transfered twice probably.
2. Version of the data to be repaired is out of control. Once
   recovery is in progress, copying data directly is danger.

The solution is dog process sends a new command to ask for repairing.
If the destinate node receives the command, tries reading the object
as normal, and then, writes to it's local storage.

Functional test case 077 is passed after patching.

Signed-off-by: Ruoyu <liangry@ucweb.com>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>