add Ruoyu as a subsystem maintainer of zookeeper components
Ruoyu is an active developer and heavy user of sheepdog. His clusters
are deployed with zookeeper so he often solves problems related to the
software. I'd like him to be a subsystem maintainer of zookeeper
components for improving our development speed.

(This patch also fixes invalid description in the file which was
introduced in a past commit)

Cc: Ruoyu <liangry@ucweb.com>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>