sheep: add epoch file recovery
If cluster members change frequently, especially the first two or
more nodes of the cluster leave and join frequently, the following
error message will be found in sheep.log

Failed to find requested tag, remote address: 127.0.0.1:7001,
op name: GET_EPOCH

This patch adds epoch recovery to reduce these errors. To achieve
it, it is better to unify epoch file because rb_node is unused
in it. And then, we can create a new epoch file as long as the file
is not found in local node.

v2 is rebased on the latest master and
1. memset's 2nd parameter, use 0 instead of '\0'
2. for loop initial declaration

Signed-off-by: Ruoyu <liangry@ucweb.com>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>