add vdi check -e to quickly check which objects are lost
Sometimes we want to quickly check whether some of the vdi objects
or data objects are lost due to unexpected issue.

Although vdi check will do, it spends a lot of time because of
too many client-server communication. And the probability of
triggering data auto fixing is quite low since the writing process
is strong consistency.

Therefore, the new option -e (--exist) check whether all the objects
related to the vdi are existed or not. It is fast because it submit
the batched object id only one time per node. I think this is enough
for the situation.

Usage: dog vdi check -e <vdiname>

Example:
$ dog vdi check -e test
test is fine, no object is missing.

$ dog vdi check -e ucweb
[127.0.0.1:7001] oid 80b8071d00000000 is missing.
[127.0.0.1:7001] oid 00b8071d000000ee is missing.
ucweb lost 2 object(s).

v4 is rebased on the latest master and

1. helping message is updated to tell user vdi check -e will
   not comparing nor repairing objects
2. the function do_obj_check is renamed as do_vdi_check_exist
3. a new command flag, SD_FLAG_CMD_FILTER, is introduced because
   both read and write is not so appropriate
4. the value of SD_FLAG_CMD_FILTER is changed because the original one
   is occupied

Signed-off-by: Ruoyu <liangry@ucweb.com>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>