sheep: snapshot and collect vdi state during joining to cluster
Newly joining sheep must know which vdi is locked or unlocked. For
propagating the state, new sheep must copy it from existing nodes
during join process.

This patch implements a solution for this. Below is the brief
description:
1. When a cluster accepts newly joining node, every existing node take
   a snapshot of vdi state at the epoch.
2. Newly joining node copy it via a newly defined request.
3. After applying the copied state to its internal state, the new
   sheep issues discard request to existing nodes.
4. The new sheep plays its operation log and solves difference between
   copied snapshot and latest state (produced by requests issued after
   taking snapshot).

With this scheme, new sheep can know which vdi is locked or unlocked.

Cc: Fabian Zimmermann <dev.faz@gmail.com>
Cc: Valerio Pachera <sirio81@gmail.com>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>