sheep/cluster: remove ephemeral node when zookeeper master leaves
The zookeeper node /sheepdog/master/000000xxxx is ephemeral, but when
the sheep process which is master in the zookeeper cluster leaves,
zookeeper will not remove the ephemeral node immediately. If the sheep
restarts at once, the sheepdog cluster will be corrupted.

This patch fixes it by trying removing the ephemeral node when
sheep leaves the cluster.

Typo for MASTER_ZNODE is also fixed in the patch.

Signed-off-by: Ruoyu <liangry@ucweb.com>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>