sheep/http: fix error in bucket_delete
In current implementation, When we create a bucket, we decide
the bnode's location in account VDI using sd_hash(bucket_name)
as key. We handle hash conflict by linear probing hash table.
Here is the bug:
When we delete a bucket, we can't discard its bnode. Because
bnode_lookup() need it to find if some bucket exists or not
by checking adjacent bnodes. Therefore, we just zero its
bnode.name when client want to delete a bucket. When we create
a bucket later, we can reuse the zero bnode if they hash to
the same location in account VDI. So, we need to change the
implementation of bnode lookup and bnode create function, too.

Signed-off-by: Bingpeng Zhu <bingpeng.zbp@alibaba-inc.com>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>