dog/node: remove checking argument in md plug/unplug since root cause is found
The purpose of the checking in the function is to ensure the
argument is not null. But it is not needed any longer because
the argument should be guaranteed in the framework as long as
CMD_NEED_ARG is specified in subcommand.

Since the problem is solved in commit e6d1706, we can remove
the checking safely.

Signed-off-by: Ruoyu <liangry@ucweb.com>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>