dog: fix cluster-wide snapshot inconsistency due to vdi deleted
Once vdi snapshot is deleted before executing 'dog cluster snapshot
save', error messsage just like as below is seen.

Failed to read object 807c2b2500000000 No object found
Failed to read inode header

Once a vdi is deleted, cluster snapshot save will not backup it to
farm so that cluster snapshot load cannot restore it to new cluster.
However, the vdi id is marked as in use for safety, vdi list
will try to read all vdis in use. Therefore, dog cannot read the object.

The solution is saving the deleted vdi object too. To achieve it,
a parameter, no_deleted, is added to the function parse_vdi.
The default value is true. But for cluster snapshot save, it is false.

Reported-by: Valerio Pachera <sirio81@gmail.com>
Cc: Liu Yuan <namei.unix@gmail.com>
Signed-off-by: Ruoyu <liangry@ucweb.com>
Signed-off-by: Liu Yuan <namei.unix@gmail.com>