dog/vdi: enhance find_vdi_name to allow returning SD_RES_XXX
Current return value of the function find_vdi_name is quite simple. -1
stands for failure and 0 stands for success. It is not enough if caller
A expects it returns information of a VDI and caller B expects it
returns the VDI is existed or not.

So, it is better to enhance it to return SD_RES_XXX.

As a result, the error message printed inside the function should be
moved to the callers.

Signed-off-by: Ruoyu <liangry@ucweb.com>
Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>